Add typescript-root-index plugin. Remove old generator
