--- packages/core/src/generate/generateAppFiles/generatorRootIndex/index.ts
@@ -1,5 +1,0 @@
-import { generateRootIndexFile } from "./generateRootIndexFile";
-
-export default {
-  generate: generateRootIndexFile,
-};
