+++ packages/core/src/plugins/typescript-root-index/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TypescriptRootIndexPlugin";
